adding UI messaging

+++ fixorphans.php
@@ -3,13 +3,26 @@ include ("header.php");
 include ("pass.php");
 $walletDir = "/home/stakebox/wallets/".$currentWallet;
 
-$updated = $_GET["updated"];
+$status = $_GET["status"];
 
-exec("killall navcoind");
-print '<h2>Restarting the navcoin daemon and flushing out txes. This can take up to 15 minutes.</h2>';
-exec("/usr/local/bin/navcoind -datadir=/home/stakebox/.navcoin4 -zapwallettxes");
+if ($status == "zapwallettxes") {
+  exec("killall navcoind");
+  exec("/usr/local/bin/navcoind -datadir=/home/stakebox/.navcoin4 -zapwallettxes");
+  print("<h2>Orphans have been fixed, wallet has restarted</h2>");
+} else { ?>
+  <h2>Fix Orphan Stakes</h2>
 
-?>
-</div>
-</div>
+  <p>If your balance is not displaying correctly it could be because some of your stakes have been orphaned.</p>
+  <p>You can try to fix the orphaned stakes by clicking the button below. Be patient, it can take up to 15 minutes for your wallet to repair.</p>
+
+  <form action="power" method="POST"><input type="hidden">
+      <div class="input-group">
+        <button class='btn btn-default' type="submit" name="status" value="zapwallettxes">Fix Orphans</button>
+      </div><!-- /input-group -->
+  </form>
+
+  </div>
+  </div>
+<?php } ?>
+
 <?php include ("footer.php"); ?>
